+++ Source/Clients/DotNET/Observation/IObservers.cs
@@ -1,22 +1,25 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Aksio.Cratis.Kernel.Observation;
+
 namespace Aksio.Cratis.Observation;
 
 /// <summary>
-/// Defines a system for working with observers.
+/// Defines a system for working with observers registered in the Kernel.
 /// </summary>
 public interface IObservers
 {
     /// <summary>
-    /// Initialize the observers system.
+    /// Get all observers for specific event types.
     /// </summary>
-    /// <returns>Awaitable task.</returns>
-    Task Initialize();
+    /// <param name="eventTypes">Collection of types representing events to get for.</param>
+    /// <returns>Collection of <see cref="ObserverInformation"/> holding all information about the observers.</returns>
+    Task<IEnumerable<ObserverInformation>> GetObserversForEventTypes(IEnumerable<Type> eventTypes);
 
     /// <summary>
-    /// Get all the handlers registered.
+    /// Get all observers.
     /// </summary>
-    /// <returns>Collection of <see cref="ObserverHandler"/>.</returns>
-    IEnumerable<ObserverHandler> Handlers { get; }
+    /// <returns>Collection of <see cref="ObserverInformation"/> holding all information about the observers.</returns>
+    Task<IEnumerable<ObserverInformation>> GetAllObservers();
 }