+++ Source/Kernel/Read/EventSequences/EventSequence.cs
@@ -7,6 +7,8 @@ using Aksio.Cratis.Events;
 using Aksio.Cratis.EventSequences;
 using Aksio.Cratis.Execution;
 using Microsoft.AspNetCore.Mvc;
+using Orleans;
+using IEventSequence = Aksio.Cratis.Kernel.Grains.EventSequences.IEventSequence;
 
 namespace Aksio.Cratis.Kernel.Read.EventSequences;
 
@@ -16,25 +18,57 @@ namespace Aksio.Cratis.Kernel.Read.EventSequences;
 [Route("/api/events/store/{microserviceId}/{tenantId}/sequence/{eventSequenceId}")]
 public class EventSequence : Controller
 {
-    readonly ProviderFor<IEventSequenceStorageProvider> _eventSequenceStorageProviderProvider;
+    readonly ProviderFor<IEventSequenceStorage> _eventSequenceStorageProviderProvider;
+    readonly IGrainFactory _grainFactory;
     readonly JsonSerializerOptions _jsonSerializerOptions;
     readonly IExecutionContextManager _executionContextManager;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="EventSequence"/> class.
     /// </summary>
-    /// <param name="eventSequenceStorageProviderProvider">Provider for <see cref="IEventSequenceStorageProvider"/>.</param>
+    /// <param name="eventSequenceStorageProviderProvider">Provider for <see cref="IEventSequenceStorage"/>.</param>
+    /// <param name="grainFactory"><see cref="IGrainFactory"/>.</param>
     /// <param name="jsonSerializerOptions"><see cref="JsonSerializerOptions"/> for serialization.</param>
     /// <param name="executionContextManager"><see cref="IExecutionContextManager"/>.</param>
     public EventSequence(
-        ProviderFor<IEventSequenceStorageProvider> eventSequenceStorageProviderProvider,
+        ProviderFor<IEventSequenceStorage> eventSequenceStorageProviderProvider,
+        IGrainFactory grainFactory,
         JsonSerializerOptions jsonSerializerOptions,
         IExecutionContextManager executionContextManager)
     {
         _eventSequenceStorageProviderProvider = eventSequenceStorageProviderProvider;
+        _grainFactory = grainFactory;
         _jsonSerializerOptions = jsonSerializerOptions;
         _executionContextManager = executionContextManager;
     }
+
+    /// <summary>
+    /// Get the head sequence number.
+    /// </summary>
+    /// <param name="eventSequenceId">Event sequence to get for.</param>
+    /// <param name="microserviceId">Microservice to get for.</param>
+    /// <param name="tenantId">Tenant to get for.</param>
+    /// <returns>The tail sequence number.</returns>
+    [HttpGet("next-sequence-number")]
+    public Task<EventSequenceNumber> Next(
+        [FromRoute] EventSequenceId eventSequenceId,
+        [FromRoute] MicroserviceId microserviceId,
+        [FromRoute] TenantId tenantId) =>
+        GetEventSequence(microserviceId, eventSequenceId, tenantId).GetNextSequenceNumber();
+
+    /// <summary>
+    /// Get the tail sequence number.
+    /// </summary>
+    /// <param name="eventSequenceId">Event sequence to get for.</param>
+    /// <param name="microserviceId">Microservice to get for.</param>
+    /// <param name="tenantId">Tenant to get for.</param>
+    /// <returns>The tail sequence number.</returns>
+    [HttpGet("tail-sequence-number")]
+    public Task<EventSequenceNumber> Tail(
+        [FromRoute] EventSequenceId eventSequenceId,
+        [FromRoute] MicroserviceId microserviceId,
+        [FromRoute] TenantId tenantId) =>
+        GetEventSequence(microserviceId, eventSequenceId, tenantId).GetTailSequenceNumber();
 
     /// <summary>
     /// Get events for a specific event sequence in a microservice for a specific tenant.
@@ -46,8 +80,8 @@ public class EventSequence : Controller
     [HttpGet]
     public async Task<IEnumerable<AppendedEventWithJsonAsContent>> FindFor(
         [FromRoute] EventSequenceId eventSequenceId,
-        [FromQuery] MicroserviceId microserviceId,
-        [FromQuery] TenantId tenantId)
+        [FromRoute] MicroserviceId microserviceId,
+        [FromRoute] TenantId tenantId)
     {
         var result = new List<AppendedEventWithJsonAsContent>();
         _executionContextManager.Establish(tenantId, CorrelationId.New(), microserviceId);
@@ -68,4 +102,7 @@ public class EventSequence : Controller
     /// <returns>A collection of <see cref="EventHistogramEntry"/>.</returns>
     [HttpGet("histogram")]
     public Task<IEnumerable<EventHistogramEntry>> Histogram(/*[FromRoute] EventSequenceId eventSequenceId*/) => Task.FromResult(Array.Empty<EventHistogramEntry>().AsEnumerable());
+
+    IEventSequence GetEventSequence(MicroserviceId microserviceId, EventSequenceId eventSequenceId, TenantId tenantId) =>
+        _grainFactory.GetGrain<IEventSequence>(eventSequenceId, keyExtension: new MicroserviceAndTenant(microserviceId, tenantId));
 }